WMS: 11866 --- update manifest.json (#407)
* WMS: 11866 - Major Update

Introduction (introduction.md)
Get Started (cloud-login.md)
Lab 1 (user-story.md)
For the screenshots that need updating I've left the following note(s).
> :warning: **Note:** `<PLACEHOLDER FOR SCREENSHOT>`
Lab 2 (tools.md)
Lab 3 (discover-and-catalog)
Aas discussed, the public bucket has been made copyable, so now somebody can just press a copy button to get the whole uri.
Whilst I was here I did quite bit of styling and formatting, and did some link formatting.
Also, there's an image, the last one of this lab that is very pixelated. I've noted it with
> :warning: **Note:** `<PIXELATED IMAGE... NEEDS UPDATING>`
Lab 4 (sync-link.md)
I have not touched at all! :smile:
Lab 5 (data-products.md)
As per the working document, this lab had to be restructured with the architecture moving from the top of the lab to the bottom, to keep the key terms consistent to be at the bottom of the lab. Eddie, I think you going to like me :wink: I managed to style Key Term the word in without having it be collapsable :wink:
Lab 6 (mktplace-subscribe.md)
Need to do some styling here (see screenshot what it look like)
Lab 7 (load-transform.md)
Set it up as per the word document. I've placed a footer at the bottom for screenshots.
@obarr please check this lab.
Lab 8 (hack-answers.md)
I just did the introduction of the lab as per the word document.
NEW Section -- Architecture (architecture.md)
Put in a new section for architecture as a lab -- as we discussed, and as per the working document.

I've also cleaned and deleted unnecessary files and folders.

* Update manifest.json

Update of manifest.json just to have Oracle Data Platform and removed AI in the workshoptitle.